pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cellsam"
version = "1.0.0"
authors = [
  { name = "CellSeek Development Team", email = "contact@cellseek.org" },
]
description = "Standalone Cell Segmentation with Segment Anything Model"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Image Processing",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
]
keywords = [
  "cell segmentation",
  "microscopy",
  "deep learning",
  "computer vision",
  "SAM",
]
dependencies = [
  "torch",
  "torchvision",
  "numpy",
  "scipy",
  "opencv-python",
  "matplotlib",
  "tqdm",
  "pathlib",
  "natsort",
  "segment-anything",
  "scikit-image",
  "pillow",
]

[tool.setuptools]
packages = ["cellsam"]
include-package-data = true

[tool.setuptools.package-data]
cellsam = ["*.md", "*.txt"]
